2008-04-01 Sancho Lerena <slerena@gmail.com>
    * pandoradb_data.sql: Alert type 0 is not 1. 1-3 are reserved. Removed 
    prediction module (nosense since it uses normal modules).

    * include/functions_db.php: Added wrapper give_db_value to get_db_value
    because A LOT of code uses old syntax and this makes too many errors.

    * include/functions.php: Fixed pagination issues with blocksize config.

    * pandoradb.sql: Change alert command lenght and added some comments
    for future implementation of alert criticity and others.
 
    * operation/incidents/incident.php: Fixed some bugs calling to old
    funcions / variables.

    * operation/events/events.php: Fixed some bugs calling to old
    funcions / variables. Removed graph and simplified header.

    * operation/agentes/estado_alertas.php: Changed get_db_value call
    and removed (Raul) button to create alert from here.

    * operation/agentes/estado_grupo.php: Implemented support for optional
    view of down modules and fired alerts.

    * operation/agentes/ver_agente.php, 
    operation/agentes/datos_agente.php: Uses of new get_db_value and
    new session id variable.
    
    * images/mod_async_string.png: Added image.

    * images/mod_async_inc.png: Added image.

    * images/mod_image_png.png: Added image.

    * images/mod_async_data.png: Added image.

    * images/mod_async_proc.png: Added image.

    * images/mod_image_jpg.png: Added image.

    * godmode/setup/setup.php: Several fixes.

    * godmode/agentes/alert_manager.php: Several changes for new alert manager. 
    Not finished. Interim commit.

    * godmode/agentes/configurar_agente.php: More changes needed to alert editor.

    * godmode/agentes/alert_manager_editor.php: New file for alert management.

    * godmode/modules/module_list.php: Some updates, removed invalid column.

    * godmode/alerts/modify_alert.php: Alert types < 4 cannot be modified (this
    includes combined, internal audit and Pandora FMS log).



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@783 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/godmode/alerts/modify_alert.php
@@ -66,9 +66,14 @@ if ( (give_acl($id_user, 0, "LM")==1)){
 			$tdcolor = "datos2";
 			$color = 1;
 		}
-		echo "<tr><td class='$tdcolor'><b><a href='index.php?sec=galertas&sec2=godmode/alerts/configure_alert&id_alerta=".$row["id_alerta"]."'>".$row["nombre"]."</a></b></td>";
-		echo "<td class='$tdcolor'>".$row["descripcion"]."</td>";
-		echo "<td class='$tdcolor' align='center'><a href='index.php?sec=gagente&sec2=godmode/alerts/modify_alert&borrar_alerta=".$row["id_alerta"]."' onClick='if (!confirm(\' ".$lang_label["are_you_sure"]."\')) return false;'><img border='0' src='images/cross.png'></a></td>";
+        if ($row[0] > 3){
+		    echo "<tr><td class='$tdcolor'><b><a href='index.php?sec=galertas&sec2=godmode/alerts/configure_alert&id_alerta=".$row["id_alerta"]."'>".$row["nombre"]."</a></b></td>";
+		    echo "<td class='$tdcolor'>".$row["descripcion"]."</td>";
+		    echo "<td class='$tdcolor' align='center'><a href='index.php?sec=gagente&sec2=godmode/alerts/modify_alert&borrar_alerta=".$row["id_alerta"]."' onClick='if (!confirm(\' ".$lang_label["are_you_sure"]."\')) return false;'><img border='0' src='images/cross.png'></a></td>";
+        } else {
+            echo "<tr><td class='$tdcolor'><b>".$row["nombre"]."</b></td>";
+            echo "<td class='$tdcolor'>".$row["descripcion"]."</td>";
+        }
 	}
 
 	echo "</tr></table>";
